fix: Move to using interface for better mocking at chaoshub
