+++ litmus-portal/graphql-server/graph/resolver.go
@@ -8,17 +8,21 @@ import (
 	"github.com/99designs/gqlgen/graphql"
 	"github.com/litmuschaos/litmus/litmus-portal/graphql-server/graph/generated"
 	"github.com/litmuschaos/litmus/litmus-portal/graphql-server/pkg/authorization"
+	"github.com/litmuschaos/litmus/litmus-portal/graphql-server/pkg/chaoshub"
 )
 
 // This file will not be regenerated automatically.
 //
 // It serves as dependency injection for your app, add any dependencies you require here.
 
-type Resolver struct{}
+type Resolver struct {
+	ChaosHubService chaoshub.Service
+}
 
 func NewConfig() generated.Config {
-
-	config := generated.Config{Resolvers: &Resolver{}}
+	config := generated.Config{Resolvers: &Resolver{
+		ChaosHubService: chaoshub.NewService(),
+	}}
 	config.Directives.Authorized = func(ctx context.Context, obj interface{}, next graphql.Resolver) (interface{}, error) {
 		token := ctx.Value(authorization.AuthKey).(string)
 		user, err := authorization.UserValidateJWT(token)
